dev_requirements.txt
# Packages needed if you'd like to test / develop on django-autocompleter
coverage
django-nose
nose
pre-commit
